requirements.txt
django==4.0.5
python-dotenv==0.20.0
python-telegram-bot==13.13
flake8==3.9.2

